requirements.txt
google-generativeai
streamlit
pandas
numpy
requests
python-dotenv
pyarrow